+++ NEWS.md
@@ -3,7 +3,7 @@
 * Updated R version requirement to >= 3.6.
 * Fixed error in `tm_g_gh_boxplot` when no facet variable is selected.
 * Added an additional UI input component to add an addition arbitrary horizontal line to `tm_g_gh_spaghettiplot`, `tm_g_gh_boxplot`, `tm_g_gh_density_distribution_plot` as well as two additional UI input components to add an additional horizontal and an additional vertical line to `tm_g_gh_correlationplot`.
-* Replaced calls to `utils.nest::check_slider_input` with `checkmate::assert_numeric`.
+* Removed dependency on `utils.nest` and replaced those function with equivalents from the `checkmate` package`.
 
 # teal.goshawk 0.1.10